Renamed model packages to be all lowercase
