+++ src/main/java/uk/gov/ch/model/emergencyauthcode/jsondatamodels/CorporateBodyAppointments.java
@@ -1,4 +1,4 @@
-package uk.gov.ch.model.emergencyauthcode.jsonDataModels;
+package uk.gov.ch.model.emergencyauthcode.jsondatamodels;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
